refactor position of utils
