+++ services/dynamic-scheduler/src/simcore_service_dynamic_scheduler/api/frontend/routes_internal_scheduler/_index.py
@@ -1,8 +1,11 @@
 from nicegui import APIRouter, ui
 
+from .._render_utils import base_page
+
 router = APIRouter()
 
 
 @router.page("/")
 async def index():
-    ui.label("PLACEHOLDER for internal scheduler UI")
+    with base_page():
+        ui.label("PLACEHOLDER for internal scheduler UI")
